requirements-dev-minimal.txt
python-xkcd
flask
threading
discordpy
wikipedia
bs4
replit-db-client